+++ internal/provider/reservation_resource_cidr.go
@@ -3,8 +3,6 @@ package provider
 import (
 	"context"
 	"fmt"
-	"strconv"
-	"strings"
 	"time"
 
 	ipamclient "terraform-provider-azureipam/ipamclient"
@@ -13,9 +11,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/boolplanmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/int32planmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/listplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/types"
@@ -23,49 +18,46 @@ import (
 
 // Ensure the implementation satisfies the expected interfaces.
 var (
-	_ resource.Resource                = &reservationResource{}
-	_ resource.ResourceWithConfigure   = &reservationResource{}
-	_ resource.ResourceWithImportState = &reservationResource{}
+	_ resource.Resource                = &reservationResourceCidr{}
+	_ resource.ResourceWithConfigure   = &reservationResourceCidr{}
+	_ resource.ResourceWithImportState = &reservationResourceCidr{}
 )
 
-// NewReservationResource is a helper function to simplify the provider implementation.
-func NewReservationResource() resource.Resource {
-	return &reservationResource{}
+// NewReservationCidrResource is a helper function to simplify the provider implementation.
+func NewReservationCidrResource() resource.Resource {
+	return &reservationResourceCidr{}
 }
 
-// reservationResourceModel maps the resource schema data.
-type reservationResourceModel struct {
-	Space         types.String      `tfsdk:"space"`
-	Blocks        types.List        `tfsdk:"blocks"`
-	Size          types.Int32       `tfsdk:"size"`
-	Description   types.String      `tfsdk:"description"`
-	ReverseSearch types.Bool        `tfsdk:"reverse_search"`
-	SmallestCidr  types.Bool        `tfsdk:"smallest_cidr"`
-	Id            types.String      `tfsdk:"id"`
-	Block         types.String      `tfsdk:"block"`
-	Cidr          types.String      `tfsdk:"cidr"`
-	CreatedBy     types.String      `tfsdk:"created_by"`
-	CreatedOn     timetypes.RFC3339 `tfsdk:"created_on"`
-	SettledBy     types.String      `tfsdk:"settled_by"`
-	SettledOn     timetypes.RFC3339 `tfsdk:"settled_on"`
-	Status        types.String      `tfsdk:"status"`
-	Tags          types.Map         `tfsdk:"tags"`
+// reservationResourceCidrModel maps the resource schema data.
+type reservationResourceCidrModel struct {
+	Space        types.String      `tfsdk:"space"`
+	Block        types.String      `tfsdk:"block"`
+	SpecificCidr types.String      `tfsdk:"specific_cidr"`
+	Description  types.String      `tfsdk:"description"`
+	Id           types.String      `tfsdk:"id"`
+	Cidr         types.String      `tfsdk:"cidr"`
+	CreatedBy    types.String      `tfsdk:"created_by"`
+	CreatedOn    timetypes.RFC3339 `tfsdk:"created_on"`
+	SettledBy    types.String      `tfsdk:"settled_by"`
+	SettledOn    timetypes.RFC3339 `tfsdk:"settled_on"`
+	Status       types.String      `tfsdk:"status"`
+	Tags         types.Map         `tfsdk:"tags"`
 }
 
-// reservationResource is the resource implementation.
-type reservationResource struct {
+// reservationResourceCidr is the resource implementation.
+type reservationResourceCidr struct {
 	client *ipamclient.Client
 }
 
 // Metadata returns the resource type name.
-func (r *reservationResource) Metadata(_ context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
-	resp.TypeName = req.ProviderTypeName + "_reservation"
+func (r *reservationResourceCidr) Metadata(_ context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_reservation_cidr"
 }
 
 // Schema defines the schema for the resource.
-func (r *reservationResource) Schema(_ context.Context, _ resource.SchemaRequest, resp *resource.SchemaResponse) {
+func (r *reservationResourceCidr) Schema(_ context.Context, _ resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
-		Description: "The reservation resource allows you to create a IPAM reservation in the specific space and list of blocks.",
+		Description: "The reservation resource allows you to create a IPAM reservation in the specific space and block with a fixed cidr.",
 		Attributes: map[string]schema.Attribute{
 			"space": schema.StringAttribute{
 				Description: "Name of the existing space in the IPAM application. Changing this forces a new resource to be created.",
@@ -74,45 +66,26 @@ func (r *reservationResource) Schema(_ context.Context, _ resource.SchemaRequest
 					stringplanmodifier.RequiresReplace(),
 				},
 			},
-			"blocks": schema.ListAttribute{
+			"block": schema.StringAttribute{
 				Description: "List with the names of blocks in the specified space in which the reservation is to be create. The list is evaluated in the order provider. Changing this forces a new resource to be created.",
 				Required:    true,
-				ElementType: types.StringType,
-				PlanModifiers: []planmodifier.List{
-					listplanmodifier.RequiresReplace(),
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
 				},
 			},
-			"size": schema.Int32Attribute{
-				Description: "Integer value to indicate the subnet mask bits, which defines the size of the vnet to reserve (example 24 for a /24 subnet). Changing this forces a new resource to be created.",
-				Required:    true,
-				PlanModifiers: []planmodifier.Int32{
-					int32planmodifier.RequiresReplace(),
+			"specific_cidr": schema.StringAttribute{
+				Description: "The specific CIDR to reserve, in cidr notation. At least one of size or specific_cidr attribute must be specified. Not allowed if more than one block is specified.",
+				Optional:    true,
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
 				},
 			},
 			"description": schema.StringAttribute{
 				Description: "Description text that describe the reservation, that will be added as an additional tag.",
 				Optional:    true,
 			},
-			"reverse_search": schema.BoolAttribute{
-				Description: "New networks will be created as close to the end of the block as possible?. Defaults to `false`. Changing this forces a new resource to be created.",
-				Optional:    true,
-				PlanModifiers: []planmodifier.Bool{
-					boolplanmodifier.RequiresReplace(),
-				},
-			},
-			"smallest_cidr": schema.BoolAttribute{
-				Description: "New networks will be created using the smallest possible available block? (e.g. it will not break up large CIDR blocks when possible).Defaults to `false`. Changing this forces a new resource to be created.",
-				Optional:    true,
-				PlanModifiers: []planmodifier.Bool{
-					boolplanmodifier.RequiresReplace(),
-				},
-			},
 			"id": schema.StringAttribute{
 				Description: "The unique identifier of the generated reservation.",
-				Computed:    true,
-			},
-			"block": schema.StringAttribute{
-				Description: "Block where the reservation have been created.",
 				Computed:    true,
 			},
 			"cidr": schema.StringAttribute{
@@ -151,26 +124,24 @@ func (r *reservationResource) Schema(_ context.Context, _ resource.SchemaRequest
 }
 
 // Create a new resource.
-func (r *reservationResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
+func (r *reservationResourceCidr) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
 	// Retrieve values from plan
-	var plan reservationResourceModel
+	var plan reservationResourceCidrModel
 	diags := req.Plan.Get(ctx, &plan)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
 	}
 
-	var blocks *[]string
-	diag:= plan.Blocks.ElementsAs(ctx, &blocks, false)
-	resp.Diagnostics.Append(diag...)
- 	reservation, err := r.client.CreateReservation(
+	block := []string{plan.Block.ValueString()}
+	reservation, err := r.client.CreateReservation(
 		plan.Space.ValueString(),
-		*blocks,
+		block,
 		plan.Description.ValueStringPointer(),
-		plan.Size.ValueInt32Pointer(),
 		nil,
-		plan.ReverseSearch.ValueBool(),
-		plan.SmallestCidr.ValueBool(),
+		plan.SpecificCidr.ValueStringPointer(),
+		false,
+		false,
 	)
 	if err != nil {
 		resp.Diagnostics.AddError(
@@ -181,8 +152,18 @@ func (r *reservationResource) Create(ctx context.Context, req resource.CreateReq
 	}
 
 	// Map response body to schema and populate Computed attribute values
-	flattenReservation(reservation, &plan)
+	flattenReservationCidr(reservation, &plan)
 	plan.Tags, _ = types.MapValueFrom(ctx, types.StringType, reservation.Tags)
+	// //Calculate requested size from assigned Cidr
+	// size, err := strconv.Atoi(strings.Split(reservation.Cidr, "/")[1])
+	// plan.Size = types.Int32Value(int32(size))
+	// if err != nil {
+	// 	resp.Diagnostics.AddError(
+	// 		"Error Reading AzureIpam Reservation",
+	// 		"Could not determinate requested size for Reservation with id "+plan.Id.ValueString()+": "+err.Error(),
+	// 	)
+	// 	return
+	// }
 
 	// Set state to fully populated data
 	diags = resp.State.Set(ctx, plan)
@@ -193,9 +174,9 @@ func (r *reservationResource) Create(ctx context.Context, req resource.CreateReq
 }
 
 // Read resource information.
-func (r *reservationResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
+func (r *reservationResourceCidr) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
 	// Get current state
-	var state reservationResourceModel
+	var state reservationResourceCidrModel
 	diags := req.State.Get(ctx, &state)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
@@ -215,20 +196,8 @@ func (r *reservationResource) Read(ctx context.Context, req resource.ReadRequest
 	}
 
 	// Overwrite items with refreshed state
-	flattenReservation(reservation, &state)
-	// state.ReverseSearch = types.BoolValue(reverse_search)
-	// state.SmallestCidr = types.BoolValue(smallest_cidr)
+	flattenReservationCidr(reservation, &state)
 	state.Tags, _ = types.MapValueFrom(ctx, types.StringType, reservation.Tags)
-	//Calculate requested size from assigned Cidr
-	size, err := strconv.Atoi(strings.Split(reservation.Cidr, "/")[1])
-	state.Size = types.Int32Value(int32(size))
-	if err != nil {
-		resp.Diagnostics.AddError(
-			"Error Reading AzureIpam Reservation",
-			"Could not determinate requested size for Reservation with id "+state.Id.ValueString()+": "+err.Error(),
-		)
-		return
-	}
 
 	// Set refreshed state
 	diags = resp.State.Set(ctx, &state)
@@ -239,8 +208,8 @@ func (r *reservationResource) Read(ctx context.Context, req resource.ReadRequest
 }
 
 // Update not allowed, returning readed plan as current state.
-func (n *reservationResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
-	var model reservationResource
+func (n *reservationResourceCidr) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+	var model reservationResourceCidr
 
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &model)...)
 	if resp.Diagnostics.HasError() {
@@ -250,9 +219,9 @@ func (n *reservationResource) Update(ctx context.Context, req resource.UpdateReq
 	resp.Diagnostics.Append(resp.State.Set(ctx, &model)...)
 }
 
-func (r *reservationResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+func (r *reservationResourceCidr) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
 	// Retrieve values from state
-	var state reservationResourceModel
+	var state reservationResourceCidrModel
 	diags := req.State.Get(ctx, &state)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
@@ -275,7 +244,7 @@ func (r *reservationResource) Delete(ctx context.Context, req resource.DeleteReq
 }
 
 // Configure adds the provider configured client to the resource.
-func (r *reservationResource) Configure(_ context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
+func (r *reservationResourceCidr) Configure(_ context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
 	// Add a nil check when handling ProviderData because Terraform
 	// sets that data after it calls the ConfigureProvider RPC.
 	if req.ProviderData == nil {
@@ -295,15 +264,16 @@ func (r *reservationResource) Configure(_ context.Context, req resource.Configur
 	r.client = client
 }
 
-func (r *reservationResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
+func (r *reservationResourceCidr) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
 	// Retrieve import ID and save to id attribute
 	resource.ImportStatePassthroughID(ctx, path.Root("id"), req, resp)
 }
 
-func flattenReservation(reservation *ipamclient.Reservation, model *reservationResourceModel) {
+func flattenReservationCidr(reservation *ipamclient.Reservation, model *reservationResourceCidrModel) {
 	model.Id = types.StringValue(reservation.Id)
 	model.Space = types.StringValue(reservation.Space)
 	model.Block = types.StringValue(reservation.Block)
+	model.SpecificCidr = types.StringValue(reservation.Cidr) //the requestd cidr is also the assigned cidr
 	model.Cidr = types.StringValue(reservation.Cidr)
 	model.Description = types.StringValue(reservation.Description)
 	model.CreatedOn = timetypes.NewRFC3339TimeValue(time.Unix(int64(reservation.CreatedOn), 0))